+++ config.py
@@ -1,0 +1,4 @@
+cell_size = 60
+gravity_acceleration = 10
+TPS = 60  # Обновления (tick) в секунду
+width, height = 1920, 1080
